// app/state/enums/eventPrompt.ts
export enum EventPrompt {
  FavSongByArtist = 'My favorite song by an artist on the lineup',
  FirstSeenArtist = 'The first time I saw an artist on the lineup',
  LastSeenArtist = 'The last time I saw an artist on the lineup',
  OneSongLive = 'The one song I want to hear live at this event',
  EventVenueFeel = 'How I feel about the venue',
}