// app/state/enums/prompt.ts
export enum Prompt {
  FavoriteArtist = 'My favorite artist is',
  FavoriteSong = 'My favorite song is',
  FavoriteRemix = 'My favorite remix is',
  FavoriteAlbum = 'My favorite album is',
  FavoriteEP = 'My favorite EP is',
  FavoriteCollab = 'My favorite collab is',
  FavoriteRecordLabel = 'My favorite record label is',
  FavoriteLyrics = 'My favorite electronic music lyrics are',
  FirstLiveExperience = 'My first live electronic music experience was',
  FavoriteThrowbackTrack = 'My favorite throwback track is',
  FavoriteFestivalMemory = 'My favorite festival memory is',
  DreamCollab = 'My dream collab is',
  DreamB2BSet = 'My dream B2B set is',
  GuiltyPleasureSong = 'My guilty pleasure song is',
  BestConcert = 'The best concert I ever attended was',
  BestFestival = 'The best festival I ever attended was',
  BestSet = 'The best set I ever saw was',
  BestB2BSet = 'The best B2B set I ever saw was',
  MostUnderratedArtist = 'The most underrated electronic music artist is',
  MostAnticipatedArtist = 'The most anticipated electronic music artist I’ve yet to see is',
  BiggestMisconceptionAboutElectronicMusic = 'The biggest misconception about electronic music is',
  PerfectSunsetTrack = 'The perfect sunset track is',
  SongToCheerUp = 'The song I listen to cheer myself up is',
  LoveMostAboutElectronicMusic = 'The thing I love most about electronic music',
  RemixBetterThanOriginal = 'A remix that is better than the original',
  RantAboutElectronicMusic = 'A quick rant about electronic music',
  DJStageName = 'If I was a DJ, my stage name would be',
  HouseOrTechno = 'House or techno?',
  DubstepOrTrap = 'Dubstep or trap?',
  RideRailOrHangBack = 'Riding rail or hanging at the back?',
  MoshOrShuffle = 'Moshing or shuffling?',
  AllIAskAtShow = 'All I ask at a show is that you',
  // ChangeMyMindAbout = 'Change my mind about',
  // GreenFlags = 'Green flags I look for',
  // RedFlags = 'Red flags I look out for',
  // DontHateMeIf = 'Don’t hate me if I',
  // WontShutUpAbout = 'I won’t shut up about',
  // ConvincedThat = 'I’m convinced that',
  // SamePageAbout = 'Let’s make sure we’re on the same page about',
  // PetPeeve = 'My biggest pet peeve is',
  // SimplePleasures = 'My simple pleasures are',
}